extract base class from image view to inherit video view from
